move src/metadata.ts to src/lib/
